requirements.txt
fastapi
uvicorn
python-whois
dnspython
requests
jinja2
python-multipart
